+++ Servers/QuazalServer/RDVServices/GameServices/PS3DriverServices/DriverUniqueIDService.cs
@@ -1,10 +1,9 @@
 using QuazalServer.QNetZ.Attributes;
 using QuazalServer.QNetZ.Interfaces;
-using QuazalServer.RDVServices.RMC;
 
 namespace QuazalServer.RDVServices.GameServices.PS3DriverServices
 {
-    [RMCService(RMCProtocolId.DriverUniqueIDService)]
+    [RMCService((ushort)RMCProtocolId.DriverUniqueIDService)]
     public class DriverUniqueIDService : RMCServiceBase
     {
         static uint UniqueIDCounter = 26434;
