+++ Servers/QuazalServer/RDVServices/GameServices/PS3DriverServices/RemoteLogDeviceService.cs
@@ -4,7 +4,7 @@ using QuazalServer.RDVServices.RMC;
 
 namespace QuazalServer.RDVServices.GameServices.PS3DriverServices
 {
-    [RMCService(RMCProtocolId.RemoteLogDeviceService)]
+    [RMCService((ushort)RMCProtocolId.RemoteLogDeviceService)]
     public class RemoteLogDeviceService : RMCServiceBase
     {
         [RMCMethod(1)]
