+++ Servers/QuazalServer/RDVServices/GameServices/PS3GhostbustersServices/AccountManagementService.cs
@@ -6,7 +6,7 @@ using QuazalServer.RDVServices.RMC;
 
 namespace QuazalServer.RDVServices.GameServices.PS3GhostbustersServices
 {
-    [RMCService(RMCProtocolId.AccountManagementService)]
+    [RMCService((ushort)RMCProtocolId.AccountManagementService)]
     public class AccountManagementService : RMCServiceBase
     {
         [RMCMethod(1)]
