+++ Servers/QuazalServer/RDVServices/GameServices/PS3UbisoftServices/LocalizationService.cs
@@ -4,7 +4,7 @@ using QuazalServer.RDVServices.RMC;
 
 namespace QuazalServer.RDVServices.GameServices.PS3UbisoftServices
 {
-    [RMCService(RMCProtocolId.LocalizationService)]
+    [RMCService((ushort)RMCProtocolId.LocalizationService)]
     public class LocalizationService : RMCServiceBase
     {
         private static string LocaleCode = "en-US";
